Address presumbit errors.
1. Remove ldap after TODO.
2. Move DISALLOW_COPY_AND_ASSIGN into private section.
3. Remove redundant blank lines.
4. Add explicit to ctors with single parameter and remove explicit for
   ctors with multiple parameters.
5. Add required STL header.
6. Replace DCHECK(a == b) with DCHECK_EQ(a, b) and do the same thing for
   similar checks.
7. Replace unsigned long long with uint64
8. Remove unnecessary ';' after '}'.
9. Add comment to end of namespace bracket.
10. Move static const into anonymous namespace.
11. Add {} to if statement with multiline body.
12. Replace C style cast with C++ style cast.

b/34275215

+++ cobalt/media/blink/resource_multibuffer_data_provider.h
@@ -51,8 +51,8 @@ class MEDIA_BLINK_EXPORT ResourceMultiBufferDataProvider
       blink::WebURLRequest& newRequest,
       const blink::WebURLResponse& redirectResponse) OVERRIDE;
   void didSendData(blink::WebURLLoader* loader,
-                   unsigned long long bytesSent,
-                   unsigned long long totalBytesToBeSent) OVERRIDE;
+                   uint64 bytesSent,
+                   uint64 totalBytesToBeSent) OVERRIDE;
   void didReceiveResponse(blink::WebURLLoader* loader,
                           const blink::WebURLResponse& response) OVERRIDE;
   void didDownloadData(blink::WebURLLoader* loader,
